requirements.txt
pandas==2.0.3
opendatasets==0.1.22
pyspark==3.3.0
pyyaml==6.0.1
pytest==7.1.2
pytest-cov==4.1.0